refactor: rename foundation/ package to core/
Phase 2 of v3.0.0 modularization:
- Rename foundation/ directory to core/
- Update all imports across codebase
- Update documentation references in ADRs and ROADMAP
- Update mock patches in tests

Breaking change for external code that imports from foundation package.
All 148 tests pass.
