+++ core/education.py
